Move the same `displayRatingPrompt` to an extension for all view controllers to call.
